package/linux-headers: add support for Linux 5.0 kernel headers
This sets 5.0 to be the default, moving from 4.19 which was used instead
of 4.20 for the LTS branch.

We need to update linux/linux.hash as part of this commit, because
package/linux-headers/linux-headers.hash is a symlink to
linux/linux.hash.

Signed-off-by: Joel Stanley <joel@jms.id.au>
--
I kept Thomas' message about the hash as a reminder. We could drop that
text.