arch/powerpc: Clarify generic CPUs
In the case where a specific CPU is not selected, set an appropriate
gcc target for the architecture.

For clarity this introduces a BR2_generic_powerpc64 as a generic 64 bit
CPU as this makes the it clearer when browsing menuconfig.

There's no such thing as a generic ppc64le CPU. Attempting to configure
glibc for powerpc64le fails:

 configure: error: ***  POWER8 or newer is required on powerpc64le.
 __builtin_signbit is broken.  GCC 7.4 or newer is required to resolve
 (PR83862). The compiler must support -mabi=ieeelongdouble and
 -mlong-double-128 simultaneously.

Hence set Power8 as the default CPU when using ppc64le.

Signed-off-by: Joel Stanley <joel@jms.id.au>