arch/powerpc: Add Power10 CPU type
The Power10 CPU was released by IBM in 2021. It is a 64-bit CPU
implementing ISA v3.1. It runs in big and little endian modes, and
supports VSX and AltiVec.

It introduces prefix (8 byte) instructions along side the existing 4
byte instructions, and new SIMD MMA instructions.

Power10 is supported as of GCC 10 and Binutils 2.35.

Signed-off-by: Joel Stanley <joel@jms.id.au>